+++ examples/fastify-ssr/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.2](https://github.com/nearform/graphql-hooks/compare/fastify-ssr@2.0.1...fastify-ssr@2.0.2) (2019-10-24)
+
+**Note:** Version bump only for package fastify-ssr
+
+
+
+
+
+## [2.0.1](https://github.com/nearform/graphql-hooks/compare/fastify-ssr@2.0.0...fastify-ssr@2.0.1) (2019-10-04)
+
+**Note:** Version bump only for package fastify-ssr
+
+
+
+
+
+# [2.0.0](https://github.com/nearform/graphql-hooks/compare/fastify-ssr@1.0.10...fastify-ssr@2.0.0) (2019-08-20)
+
+**Note:** Version bump only for package fastify-ssr
+
+
+
+
+
+## [1.0.10](https://github.com/nearform/graphql-hooks/compare/fastify-ssr@1.0.9...fastify-ssr@1.0.10) (2019-08-05)
+
+**Note:** Version bump only for package fastify-ssr
+
+
+
+
+
 ## [1.0.9](https://github.com/nearform/graphql-hooks/compare/fastify-ssr@1.0.8...fastify-ssr@1.0.9) (2019-06-21)