+++ packages/graphql-hooks/CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.2](https://github.com/nearform/graphql-hooks/compare/graphql-hooks@4.0.1...graphql-hooks@4.0.2) (2019-10-24)
+
+
+### Bug Fixes
+
+* publishing issues ([#389](https://github.com/nearform/graphql-hooks/issues/389)) ([7e1c4a2](https://github.com/nearform/graphql-hooks/commit/7e1c4a2ba8a6c08d09139733b2b897b81374fbac))
+* Server-side rendering ([#382](https://github.com/nearform/graphql-hooks/issues/382)) ([7256bde](https://github.com/nearform/graphql-hooks/commit/7256bde4e52fc78479c887c7671eb7fb82cbd0d0)), closes [#371](https://github.com/nearform/graphql-hooks/issues/371)
+
+
+
+
+
+## [4.0.1](https://github.com/nearform/graphql-hooks/compare/graphql-hooks@4.0.0...graphql-hooks@4.0.1) (2019-10-04)
+
+
+### Bug Fixes
+
+* Add support for generic Variables TS type fixes [#348](https://github.com/nearform/graphql-hooks/issues/348) ([#353](https://github.com/nearform/graphql-hooks/issues/353)) ([4c91939](https://github.com/nearform/graphql-hooks/commit/4c91939))
+
+
+
+
+
+# [4.0.0](https://github.com/nearform/graphql-hooks/compare/graphql-hooks@3.6.1...graphql-hooks@4.0.0) (2019-08-20)
+
+**Note:** Version bump only for package graphql-hooks
+
+
+
+
+
+## [3.6.1](https://github.com/nearform/graphql-hooks/compare/graphql-hooks@3.6.0...graphql-hooks@3.6.1) (2019-08-05)
+
+**Note:** Version bump only for package graphql-hooks
+
+
+
+
+
 # [3.6.0](https://github.com/nearform/graphql-hooks/compare/graphql-hooks@3.5.1...graphql-hooks@3.6.0) (2019-06-21)